assets: refactor search components to allow re-usability
