+++ invenio_communities/assets/semantic-ui/js/invenio_communities/community/index.js
@@ -1,0 +1,7 @@
+export { default as CarouselItem } from "./communitiesCarousel/CarouselItem";
+export { default as CommunitiesCarousel } from "./communitiesCarousel/CommunitiesCarousel";
+export { default as FeaturedCommunities } from "./featuredCommunities/FeaturedCommunities";
+export { default as FeaturedCommunity } from "./featuredCommunities/FeaturedCommunity";
+export { CommunityItem } from "./communitiesItems/CommunityItem";
+export { CommunityCompactItem } from "./communitiesItems/CommunityCompactItem";
+export * from "./searchComponents";
