+++ invenio_communities/assets/semantic-ui/js/invenio_communities/community/searchComponents/CommunitiesSearchBarElement.js
@@ -1,0 +1,46 @@
+import { i18next } from "@translations/invenio_communities/i18next";
+import React from "react";
+import { withState } from "react-searchkit";
+import { Input } from "semantic-ui-react";
+
+export const CommunitiesSearchBarElement = withState(
+  ({
+    placeholder: passedPlaceholder,
+    queryString,
+    onInputChange,
+    updateQueryState,
+    currentQueryState,
+  }) => {
+    const placeholder = passedPlaceholder || i18next.t("Search");
+
+    const onSearch = () => {
+      updateQueryState({ ...currentQueryState, queryString });
+    };
+    const onBtnSearchClick = () => {
+      onSearch();
+    };
+    const onKeyPress = (event) => {
+      if (event.key === "Enter") {
+        onSearch();
+      }
+    };
+
+    return (
+      <Input
+        action={{
+          icon: "search",
+          onClick: onBtnSearchClick,
+          className: "search",
+        }}
+        fluid
+        placeholder={placeholder}
+        onChange={(event, { value }) => {
+          onInputChange(value);
+        }}
+        value={queryString}
+        onKeyPress={onKeyPress}
+      />
+    );
+  }
+);
+
