+++ invenio_communities/assets/semantic-ui/js/invenio_communities/community/searchComponents/index.js
@@ -1,0 +1,5 @@
+export { CommunitiesResults } from "./CommunitiesResults";
+export { CommunitiesSearchLayout } from "./CommunitiesSeachLayout";
+export { ResultsGridItemTemplate } from "./ResultsGridItemTemplate";
+export { CommunitiesSearchBarElement } from "./CommunitiesSearchBarElement";
+export { CommunitiesEmptySearchResults } from "./CommunitiesEmptySearchResults";
